requirements.txt
aiohttp==3.8.3
argparse==1.4.0
black==23.1.0
python-dateutil==2.8.2